fix: offline support tests break randomly as they run parallel by default (#2046)
* chore: replace dropping tables from better sqlite to using quick sqlite client directly

* chore: wip

* Revert "chore: temporarily disable offline support tests"

This reverts commit 655be7864d5095d07ca88234f464ca4cc3ed7fd3.

* fix: run offline support tests sequentilly

* fix: do not do tsc on the offline support files
